requirements.txt
# Core dependencies
hydra-core>=1.3.2
numpy==1.26.4
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
sentence-transformers>=2.7.0
faiss-cpu==1.7.2
pyserini>=0.24.0
rank-bm25>=0.2.2
torch>=2.1
transformers>=4.40
httpx>=0.27.0
structlog>=23.1
loguru>=0.7.0
mlflow>=2.11
prefect>=2.16
litellm>=1.37.16
pydantic>=2.6
python-dotenv>=1.0
tqdm>=4.66
nltk>=3.8
orjson>=3.10
PyYAML>=6.0
joblib>=1.3
einops>=0.6
uvicorn>=0.23
fastapi>=0.110

# Optional dependencies (for GPU support)
# Uncomment the following line if using GPU
# faiss-gpu==1.7.2

# Development dependencies
pytest>=7.0
black>=23.0
flake8>=6.0